Use output resource to extract terraform templates
Use a variable output resource and extract it from makefile to the right
file Instead of using the hack of calling  `local-exec` provisioner in the
terraform template resource, which calls bash to create the external
file.

This is more elegant and solves a pair of issues: If the file is missing
it will be always created and it will avoid bash parsing and removing
single quotes in the file.